Refactor stories' controller to own package
