+++ controller/stories/stories.go
@@ -1,4 +1,4 @@
-package controller
+package stories
 
 import (
 	"encoding/json"
@@ -8,13 +8,14 @@ import (
 
 	"github.com/go-chi/chi/v5"
 
+	"github.com/source-academy/stories-backend/controller"
 	"github.com/source-academy/stories-backend/model"
 	storyviews "github.com/source-academy/stories-backend/view/stories"
 )
 
 func GetStories(w http.ResponseWriter, r *http.Request) {
 	stories := model.GetAllStories()
-	EncodeJSONResponse(w, stories)
+	controller.EncodeJSONResponse(w, stories)
 }
 
 func GetStory(w http.ResponseWriter, r *http.Request) {
@@ -25,7 +26,7 @@ func GetStory(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	story := model.GetStoryByID(storyID)
-	EncodeJSONResponse(w, story)
+	controller.EncodeJSONResponse(w, story)
 }
 
 func CreateStory(w http.ResponseWriter, r *http.Request) {
@@ -36,6 +37,6 @@ func CreateStory(w http.ResponseWriter, r *http.Request) {
 	}
 
 	model.CreateStory(story)
-	EncodeJSONResponse(w, &story)
+	controller.EncodeJSONResponse(w, &story)
 	w.WriteHeader(http.StatusCreated)
 }
